fixed hardcoded ex

+++ C03/ex01/ft_strncmp.c
@@ -6,18 +6,18 @@
 /*   By: sezequie <sezequie@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/06 04:40:37 by sezequie          #+#    #+#             */
-/*   Updated: 2023/09/13 18:24:01 by sezequie         ###   ########.fr       */
+/*   Updated: 2023/09/17 18:10:18 by sezequie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 int	ft_strncmp(char *s1, char *s2, unsigned int n)
 {
-	int				i;
+	unsigned int	i;
 	unsigned char	c1;
 	unsigned char	c2;
 
 	i = 0;
-	while ((s1[i] != '\0' || s2[i] != '\0') && (unsigned int)i < n)
+	while ((s1[i] != '\0' || s2[i] != '\0') && i < n)
 	{
 		c1 = s1[i];
 		c2 = s2[i];
